local/requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0

# MLflow for experiment tracking
mlflow>=2.8.0

# Data handling and visualization
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# For local development
jupyter>=1.0.0
